pyproject.toml
[project]
name = "articlegenerate"
version = "0.1.0"
description = ""
authors = [
    {name = "algernon"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "requests (>=2.32.5,<3.0.0)",
    "websocket-client (>=1.9.0,<2.0.0)",
    "flask (>=3.1.2,<4.0.0)",
    "ffmpeg-python (>=0.2.0,<0.3.0)",
    "edge-tts (>=7.2.7,<8.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
